Namespace Authorisation class
The Authorisation class is a publisher dfe sign in authorisation.

Namespacing it accordingly for clarity.
