+++ app/services/publishers/dfe_sign_in/authorisation.rb
@@ -1,0 +1,34 @@
+require "dsi_client"
+
+module Publishers::DfeSignIn
+  class Authorisation
+    ROLES = {
+      publisher: ENV.fetch("DFE_SIGN_IN_PUBLISHER_ROLE_ID"),
+      support_user: ENV.fetch("DFE_SIGN_IN_SUPPORT_USER_ROLE_ID"),
+    }.freeze
+
+    def initialize(dsi_client: nil, **)
+      @dsi_client = dsi_client || DSIClient.new(**)
+    end
+
+    def authorised_publisher?
+      authorised?(:publisher)
+    end
+
+    def authorised_support_user?
+      authorised?(:support_user)
+    end
+
+    private
+
+    def authorised?(role)
+      @dsi_client.role_ids.include?(ROLES[role])
+    rescue DSIClient::RequestInvalid
+      false
+    rescue DSIClient::RequestFailed => e
+      raise ExternalServerError, e.message
+    end
+
+    class ExternalServerError < StandardError; end
+  end
+end
