+++ spec/services/publishers/dfe_sign_in/authorisation_spec.rb
@@ -1,6 +1,7 @@
 require "rails_helper"
+require "dsi_client"
 
-RSpec.describe Authorisation do
+RSpec.describe Publishers::DfeSignIn::Authorisation do
   subject(:authorisation) do
     described_class.new(organisation_id: "123", user_id: "456", dsi_client:)
   end
@@ -11,21 +12,21 @@ RSpec.describe Authorisation do
 
   describe "#authorised_publisher?" do
     context "when roles include the publisher role ID" do
-      let(:role_ids) { ["test-publisher-role-id"] }
+      let(:role_ids) { %w[test-publisher-role-id] }
 
-      it { should be_authorised_publisher }
+      it { is_expected.to be_authorised_publisher }
     end
 
     context "when roles do not include the publisher role ID" do
-      let(:role_ids) { ["unknown-role-id"] }
+      let(:role_ids) { %w[unknown-role-id] }
 
-      it { should_not be_authorised_publisher }
+      it { is_expected.not_to be_authorised_publisher }
     end
 
     context "when there are no roles" do
       let(:role_ids) { [] }
 
-      it { should_not be_authorised_publisher }
+      it { is_expected.not_to be_authorised_publisher }
     end
 
     context "when the DSI request 404s" do
@@ -34,7 +35,7 @@ RSpec.describe Authorisation do
           .and_raise(DSIClient::RequestInvalid, "404")
       end
 
-      it { should_not be_authorised_publisher }
+      it { is_expected.not_to be_authorised_publisher }
     end
 
     context "when the DSI request 500s" do
@@ -51,21 +52,21 @@ RSpec.describe Authorisation do
 
   describe "#authorised_support_user?" do
     context "when roles include the support user role ID" do
-      let(:role_ids) { ["test-support-user-role-id"] }
+      let(:role_ids) { %w[test-support-user-role-id] }
 
-      it { should be_authorised_support_user }
+      it { is_expected.to be_authorised_support_user }
     end
 
     context "when roles do not include the publisher role ID" do
-      let(:role_ids) { ["unknown-role-id"] }
+      let(:role_ids) { %w[unknown-role-id] }
 
-      it { should_not be_authorised_support_user }
+      it { is_expected.not_to be_authorised_support_user }
     end
 
     context "when there are no roles" do
       let(:role_ids) { [] }
 
-      it { should_not be_authorised_support_user }
+      it { is_expected.not_to be_authorised_support_user }
     end
 
     context "when the DSI request 404s" do
@@ -74,7 +75,7 @@ RSpec.describe Authorisation do
           .and_raise(DSIClient::RequestInvalid, "404")
       end
 
-      it { should_not be_authorised_support_user }
+      it { is_expected.not_to be_authorised_support_user }
     end
 
     context "when the DSI request 500s" do
